architecture-kafka-common/src/main/java/com/acme/architecture/common/serializer/GenericSerializer.java
package com.acme.architecture.common.serializer;

import java.io.Serializable;
import java.util.Map;

import org.apache.commons.lang3.SerializationUtils;
import org.apache.kafka.common.serialization.Serializer;

public class GenericSerializer <T extends Serializable> implements Serializer<T>  {

	@Override
	public void close() {
		
	}

	@Override
	public void configure(Map<String, ?> configs, boolean isKey) {
	
	}

	@Override
	public byte[] serialize(String topic, T data) {
		return SerializationUtils.serialize(data);
	}

}
